Extract useGlobeProjection and clean up minor style issues
Collapse the three inline useMemo blocks in MapGlobe.tsx (worldFeatures,
stateFeatures, and the projection memo) into a single useGlobeProjection
hook. Simplify GlobePin's onActivate signature to drop the unused event
argument. Fix the stylistically awkward `+ -dy` expression in
usePointerInteraction to `- dy`.
